Use TabView for wallet instead of page controller.
